feat: fixed initial skill zones after logout
This is weird and I assume it's because of Doozy UI. The problem was that when starting play mode in the editor, the pre existing skill zones were active and enabled before the skill bar's Start() function (I realized this due to the Awake function of the skill zones being called before the skill bar's Start fuintion), but when logging out and thus restarting the scene, for some reson the Start function of the skill bar's Start function was called before the skill zones (the skill zones' Awake was called after the Start function of the skill bar, but still within the same frame). This lead to a zone.isActiveAndEnabled check in the skill bar to fail and thus not adding the preexisting skill zones to the skill bar. I don't exactly know why that happened but I assume that Doozy stores some information inbetween scene loads or something.